Split DataFile into DataFile and DeleteFile.
