+++ api/src/test/java/org/apache/iceberg/TestHelpers.java
@@ -312,6 +312,11 @@ public class TestHelpers {
     }
 
     @Override
+    public FileContent content() {
+      return FileContent.DATA;
+    }
+
+    @Override
     public CharSequence path() {
       return path;
     }
